Move `FeedViewController` to production target
