+++ EssentialFeediOSTests/FeedViewControllerTests.swift
@@ -6,44 +6,7 @@
 import XCTest
 import UIKit
 import EssentialFeed
-
-final class FeedViewController: UITableViewController {
-    private var loader: FeedLoader?
-    
-    private var onViewIsAppearing: ((FeedViewController) -> Void)?
-    
-    convenience init(loader: FeedLoader) {
-        self.init()
-        self.loader = loader
-    }
-    
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        
-        refreshControl = UIRefreshControl()
-        refreshControl?.addTarget(self, action: #selector(load), for: .valueChanged)
-        
-        // Note: Using `onViewIsAppearing` to defer `beginRefreshing()` until the view is fully visible.
-        // This ensures the spinner appears correctly, addressing a change in behavior introduced in iOS 17.
-        onViewIsAppearing = { vc in
-            vc.load()
-            vc.onViewIsAppearing = nil
-        }
-    }
-    
-    public override func viewIsAppearing(_ animated: Bool) {
-        super.viewIsAppearing(animated)
-        
-        onViewIsAppearing?(self)
-    }
-    
-    @objc private func load() {
-        refreshControl?.beginRefreshing()
-        loader?.load { [weak self] _ in
-            self?.refreshControl?.endRefreshing()
-        }
-    }
-}
+import EssentialFeediOS
 
 final class FeedViewControllerTests: XCTestCase {
